[#3355] - move `ArrDupe` function to a util file
