+++ Dashboard/app/js/lib/views/devices/assignment-edit-views.jsx
@@ -7,26 +7,6 @@ import observe from '../../mixins/observe';
 require('akvo-flow/views/react-component');
 
 // utils
-FLOW.ArrNoDupe = function(a) {
-  let gotIt;
-  const templ = {};
-  const tempa = Ember.A([]);
-  for (let i = 0; i < a.length; i++) {
-    templ[a.objectAt(i).clientId] = true;
-  }
-  const keys = Object.keys(templ);
-  for (let j = 0; j < keys.length; j++) {
-    gotIt = false;
-    for (let i = 0; i < a.length; i++) {
-      if (a.objectAt(i).clientId == keys[j] && !gotIt) {
-        tempa.pushObject(a.objectAt(i));
-        gotIt = true;
-      }
-    }
-  }
-  return tempa;
-};
-
 FLOW.formatDate = function(value) {
   if (!Ember.none(value)) {
     return `${value.getFullYear()}/${value.getMonth() + 1}/${value.getDate()}`;
